requirements.txt
<<<<<<< HEAD
# Streamlit Cloud Optimized Requirements
# Minimal dependencies for cloud deployment

streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
=======
# 🎯 Universal 6-Chart Trading System - PRODUCTION Requirements
# Streamlit-based application for any trader's setup
# Enhanced with Harrison's dashboard and NinjaTrader integration
# Python 3.8+ required

# Core Streamlit framework
streamlit>=1.48.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0

# OCR and Image Processing (for AlgoBox integration)
easyocr>=1.7.0
pytesseract>=0.3.10
opencv-python>=4.8.0
mss>=9.0.1
Pillow>=10.0.0

# Data handling and visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# System Monitoring & NinjaTrader Integration
psutil>=5.9.0
watchdog>=3.0.0

# HTTP & API Support
requests>=2.31.0
aiohttp>=3.9.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Web connectivity and APIs
requests>=2.31.0
websocket-client>=1.6.0
websockets>=11.0.0
aiofiles>=23.0.0

# Authentication and Security
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database
aiosqlite>=0.19.0
sqlalchemy>=2.0.0

# Scientific Computing
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
httpx>=0.25.0
requests>=2.28.0
psutil>=5.9.0

# Development and Testing (optional for cloud deployment)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Note: Some dependencies are excluded for Streamlit Cloud compatibility:
# - selenium, webdriver-manager (browser automation not needed)
# - pywin32, winsound (Windows-specific, not available on cloud)
# - pythonnet (Windows/.NET specific)
# - asyncio-mqtt (may cause issues in cloud environment)
>>>>>>> 5a0b16a5c5b4c799cab7ce6951a5422d9b89da3b
